@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 import io
-import os
 import shutil
 import subprocess
 import sys
@@ -25,16 +24,12 @@ __all__ = ["generate_barcode", "TreepoemError", "BarcodeType", "barcode_types"]
 # which disables file operations in the PS code.
 @lru_cache(maxsize=None)
 def load_bwipp() -> str:
-    if sys.version_info >= (3, 9):
-        with resources.files("treepoem").joinpath(
-            "postscriptbarcode/barcode.ps"
-        ).open() as fp:
-            return fp.read()
-    else:
-        base_dir = os.path.normpath(os.path.abspath(os.path.dirname(__file__)))
-        bwipp_path = os.path.join(base_dir, "postscriptbarcode", "barcode.ps")
-        with open(bwipp_path) as fp:
-            return fp.read()
+    with (
+        resources.files("treepoem")
+        .joinpath("postscriptbarcode/barcode.ps")
+        .open() as fp
+    ):
+        return fp.read()
 
 
 # Error handling from:
